Removing files, renaming vars, coderefactor
