+++ l2env/l2env.go
@@ -1,8 +1,6 @@
-package l2envpackege
+package l2env
 
 import (
-	"encoding/json"
-	"fmt"
 	"os"
 
 	"github.com/HexmosTech/lama2/preprocess"
@@ -10,7 +8,7 @@ import (
 	"github.com/rs/zerolog/log"
 )
 
-func ProcessEnvironmentVariables(relevantSearchString, directory string) interface{} {
+func ProcessEnvironmentVariables(searchQuery, directory string) interface{} {
 	envMap, err := preprocess.GetL2EnvVariables(directory)
 	if err != nil {
 		// Potential Errors:
@@ -18,22 +16,19 @@ func ProcessEnvironmentVariables(relevantSearchString, directory string) interfa
 		log.Error().Str("Type", "Preprocess").Msg(err.Error())
 		os.Exit(0)
 	}
-	if relevantSearchString == "" { // Handle empty -e=''
+	if searchQuery == "" { // Handle empty -e=''
 		return envMap
 	} else { // Handle non-empty -e
-		relevantEnvs := GetRelevantEnvs(envMap, relevantSearchString)
-		return relevantEnvs
+		return GetRelevantEnvs(envMap, searchQuery)
 	}
-
 }
 
-func GetRelevantEnvs(envMap map[string]map[string]interface{}, relevantSearchString string) map[string]interface{} {
+func GetRelevantEnvs(envMap map[string]map[string]interface{}, searchQuery string) map[string]interface{} {
 	envTrie := trie.New()
 	for key := range envMap {
 		envTrie.Insert(key)
 	}
 
-	searchQuery := relevantSearchString
 	suggestions := envTrie.SearchAll(searchQuery)
 	filteredEnvs := make(map[string]interface{})
 	for _, suggestion := range suggestions {
@@ -42,13 +37,4 @@ func GetRelevantEnvs(envMap map[string]map[string]interface{}, relevantSearchStr
 		}
 	}
 	return filteredEnvs
-}
-
-func createJsonENv(data interface{}) []byte {
-	filteredJSON, err := json.MarshalIndent(data, "", "  ")
-	if err != nil {
-		log.Error().Str("Type", "Preprocess").Msg(fmt.Sprintf("Failed to marshal JSON: %v", err))
-		os.Exit(0)
-	}
-	return filteredJSON
 }
