--- run.sh
@@ -1,5 +1,0 @@
-clear
-make buildme
-chmod +x ./build/l2
-rm -rf /home/lovestaco/repos/Lama2Code/l2
-cp ./build/l2 /home/lovestaco/repos/Lama2Code/
